+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * bug fixes:
   * only show relevant modalities for plole-only stops if plole also appears in stops with stop position
+  * make additional platform-only plole if pole does not cover all modalities of a platform
 
 ## Release 2025-03-09